requirements.txt
arrow>=1.3.0
beautifulsoup4>=4.12.2
binaryornot>=0.4.4
botcity-email-plugin>=0.1.5
botcity-excel-plugin>=0.1.7
botcity-framework-base>=0.4.3
botcity-framework-core>=0.4.0
botcity-framework-web>=0.8.4
botcity-maestro-sdk>=0.4.1
botcity-recorder-plugin>=0.1.3
botcity-utils>=0.1.2
cachetools>=5.3.2
certifi>=2023.11.17
cffi>=1.16.0
chardet>=5.2.0
charset-normalizer>=3.3.2
click>=8.1.7
colorama>=0.4.6
comtypes>=1.2.0
cookiecutter>=2.5.0
cryptography>=41.0.7
distro>=1.9.0
EasyProcess>=1.1
entrypoint2>=1.1
et-xmlfile>=1.1.0
google-api-core>=2.15.0
google-api-python-client>=2.111.0
google-auth>=2.25.2
google-auth-httplib2>=0.2.0
googleapis-common-protos>=1.62.0
httplib2>=0.22.0
idna>=3.6
imap-tools>=1.5.0
importlib-metadata>=7.0.1
Jinja2>=3.1.2
markdown-it-py>=3.0.0
MarkupSafe>=2.1.3
mdurl>=0.1.2
MouseInfo>=0.1.3
msedge-selenium-tools>=3.141.3
numpy>=1.26.2
oauth2client>=4.1.3
opencv-python>=             4.9.0.80
openpyxl>=3.1.2
packaging>=23.2
pandas>=2.1.4
pdfminer.six>=20221105
pdfplumber>=0.10.3
pillow>=10.2.0
pip>=23.3.2
protobuf>=4.25.1
pyasn1>=0.5.1
pyasn1-modules>=0.3.0
PyAutoGUI>=0.9.53
pycparser>=2.21
PyDrive>=1.3.1
PyGetWindow>=0.0.9
Pygments>=2.17.2
PyMsgBox>=1.0.9
PyMuPDF>=1.23.8
PyMuPDFb>=1.23.7
pyodbc>=5.0.1
pyparsing>=3.1.1
pypdfium2>=4.25.0
pyperclip>=1.8.2
PyRect>=0.2.0
PyScreeze>=0.1.27
python-dateutil>=2.8.2
python-dotenv>=1.0.0
python-slugify>=8.0.1
python-xlib>=0.33
pytweening>=1.0.7
pytz>=2023.3.post1
pyunpack>=0.3
pywin32>=306
pywinauto>=0.6.8
PyYAML>=6.0.1
regex>=2023.12.25
requests>=2.31.0
requests-toolbelt>=1.0.0
rich>=13.7.0
rsa>=4.9
selenium>=3.141.0
six>=1.16.0
soupsieve>=2.5
text-unidecode>=1.3
types-python-dateutil>=2.8.19.14
tzdata>=2023.4
Unidecode>=1.3.7
uritemplate>=4.1.1
urllib3>=1.26.18
webdriver-manager>=4.0.1
zipp>=3.17.0